requirements.txt
beautifulsoup4==4.5.3
cssutils==1.0.2
decorator==4.0.11
Django==1.11.1
django-debug-toolbar==1.8
django-extensions==1.7.9
django-filter==1.0.2
djangorestframework==3.6.2
drf-nested-routers==0.90.0
flake8==3.3.0
inflect==0.2.5
ipython==6.0.0
ipython-genutils==0.2.0
jedi==0.10.2
jellyfish==0.5.6
lxml==3.7.3
mccabe==0.6.1
mysqlclient==1.3.10
odfpy==1.3.4
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pycodestyle==2.3.1
pyexcel-io==0.3.3
pyexcel-ods==0.3.3
pyflakes==1.5.0
Pygments==2.2.0
python-dateutil==2.6.0
python-utils==2.1.0
python3-memcached==1.51
pytz==2017.2
requests==2.13.0
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
tqdm==4.11.2
traitlets==4.3.2
Unidecode==0.4.20
wcwidth==0.1.7
